requirements.txt
# Core dependencies for Portfolio Tracker
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
asyncpg>=0.27.0
python-socketio>=5.7.0
websockets>=10.4
requests>=2.28.0
python-dotenv>=0.19.0
pydantic>=1.10.0
sqlalchemy>=1.4.0
alembic>=1.8.0

# Database drivers
aiosqlite>=0.17.0
psycopg2-binary>=2.9.0

# Cryptography and security
cryptography>=3.4.0
bcrypt>=3.2.0
PyJWT>=2.4.0

# Data processing and analysis
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
plotly>=5.10.0

# Testing framework
pytest>=7.1.0
pytest-asyncio>=0.19.0
pytest-cov>=3.0.0
pytest-mock>=3.8.0
pytest-xdist>=2.5.0
pytest-html>=3.1.0
pytest-benchmark>=3.4.0
coverage>=6.4.0

# Development tools
black>=22.6.0
flake8>=5.0.0
mypy>=0.971
pre-commit>=2.20.0
isort>=5.10.0

# Documentation
sphinx>=5.1.0
sphinx-rtd-theme>=1.0.0

# Utilities
click>=8.1.0
rich>=12.5.0
tqdm>=4.64.0
python-dateutil>=2.8.0
pytz>=2022.1

# Optional: For enhanced performance
uvloop>=0.16.0; sys_platform != "win32"